fix counter operations
Operations like `enbale`/`disable`/`clear_count` may act on the entire
counter group if the third argument (0) of `ioctl` is not explicitly
specified. This is probably because the third argument of the system
call register is not cleared and the kernel treats it as
`PERF_IOC_FLAG_GROUP`.

Since `syscall::ioctl` is no longer used, it was removed to satisfy
clippy.